fix: move component hook, remove hr for custom border
